switched to float source array

+++ marching-squares/MarchingSquares.h
@@ -15,12 +15,12 @@ enum
 
 struct Source
 {
-	std::array<int, (size_t)TOTAL> arr;
+	std::array<float, (size_t)TOTAL> arr;
 };
 
 void march_squares(const Source& src);
 void draw_points(const Source& src);
-int random(int, int);
-std::array<int, (size_t)TOTAL> gen_source();
+float random(float, float);
+std::array<float, (size_t)TOTAL> gen_source();
 [[nodiscard]] int x(int);
 [[nodiscard]] int y(int);
